06-streaming/homework/pyproject.toml
[project]
name = "homework"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "apache-flink>=1.20.1",
    "ipykernel>=6.29.5",
    "kafka-python>=2.1.1",
    "setuptools>=76.0.0",
]
